@@ -5,7 +5,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.6.2</version>
+        <version>3.0.5</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>com.example</groupId>
@@ -14,7 +14,7 @@
     <name>storage-util</name>
     <description>storage-util</description>
     <properties>
-        <java.version>13</java.version>
+        <java.version>17</java.version>
         <maven.checkstyle.plugin.version>3.1.2</maven.checkstyle.plugin.version>
         <jacoco.version>0.8.7</jacoco.version>
         <minio.version>8.3.4</minio.version>
